sched: fair: Drop read_mostly attribute from sched up/downmigrate
Qualcomm considered sched_upmigrate and sched_downmigrate
to not be written to often when designing HMP.

Since the sched boosting is damaged on this kernel,
we are changing the sched_upmigrate sched_downmigrate
when our boosts get triggered fairly often.

This patch drops the read_mostly attribute from
these variables to reflect the change in handling.

Signed-off-by: Alex Naidis <alex.naidis@linux.com>